[refactor] Move components from ./src/components/common to ./src/components.
