+++ src/components/NavigatorComponent.js
