+++ src/components/Text.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { Text } from 'react-native';
 import { MediaQueryStyleSheet } from 'react-native-responsive';
 
-import Colors from '../../global/Colors';
+import Colors from '../global/Colors';
 
 export default class _Text extends React.Component {
