+++ src/components/WalletCard/styles.js
