+++ src/screens/Key/Create/CreateKeyStep1/index.js
@@ -14,7 +14,7 @@ import styles from './styles';
 import Colors from '../../../../global/Colors';
 import Header from '../../../../components/Header';
 import Screens from '../../../../global/Screens';
-import Button from '../../../../components/common/Button';
+import Button from '../../../../components/Button';
 
 class CreateKeyStep1 extends Component {
